iio: adc: ad9361: add basic support for external band control
This code is optional and enabled via the CONFIG_AD9361_EXT_BAND_CONTROL
Kconfig option.
When the RX/TX LO frequency is changed, this code gets called to apply GPIO
configuration changes, according to a list of settings defined in the
device-tree.

The essential gist of it, is that in a `ad963x` DT definition the GPIOs
that are to be used for band control will be defined like this:

	adi,band-ctl-0-gpio = <&gpio0 0 0>;
	adi,band-ctl-1-gpio = <&gpio0 1 0>;
	...
	adi,band-ctl-N-gpio = <&gpio0 N 0>;

Up to 256 GPIOs may be defined.

The numbers for `adi,band-ctl-X-gpio` do have to be incremental.
The GPIOs don't need to be incremental.

Then, in the same `ad963x` DT definition add as many settings as needed:

	adi_rx_ext_band_setting_0 {
		adi,lo-freq-min = /bits/ 64 <0>;
		adi,lo-freq-max = /bits/ 64 <5000000>;
		adi,gpio-settings = <0 ... 0>;
	};

Where:
 * `adi_rx_ext_band_setting_X` has to be incremental
 * `adi_tx_ext_band_setting_X` is the equivalent for TX band control
 * `adi,lo-freq-min` and `adi,lo-freq-max` frequency range for which to
   apply this setting
 * `adi,gpio-settings` GPIO values/operations for this setting; the
    number of values in this u32 array must match the number of
    `adi,band-ctrl-X-gpio` properties
    Values/ops are:
     - 0 - nop
     - 1 - set as output-low
     - 2 - set as output-high
     - 3 - set as input [hi-z]

This change is just the basic functionality, and only handles settings for
GPIOs and switching between them.
Sequencing and init/uninit values will be added in subsequent commits.

Not specifying a `adi,gpio-settings` property (or an empty one) is
valid and is regarded as a NOP (no-operation).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>