iio: adc: ad9361: move RX/TX port selection on state struct
The TX/RX port selection can be controlled via sysfs, so it makes sense to
make it a member of the state struct [vs the platform data struct].

Added accessor functions to update the selections, and to validate the
input.
The accessor functions will be used later in the band control.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>